extensions/extensions-data-delivery/aissemble-extensions-data-delivery-spark-py/pyproject.toml
[tool.poetry]
name = "aissemble-extensions-data-delivery-spark-py"
version = "1.7.0.dev"
description = "Contains the core Python functionality of data delivery for Spark"
authors = ["aiSSEMBLE Baseline Community <aissemble@bah.com>"]
readme = "README.md"
packages = [
    {include = "data_delivery_spark_py", from="src"},
]

[tool.poetry.dependencies]
python = ">=3.11.4, <4"
krausening = ">=19"
pyspark = "3.4.0"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
black = ">=22.1.0"
behave = ">=1.2.6"
nose = ">=1.3.7"
kappa-maki = ">=1.0.1"


[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"

